Added interactable and colored logic references to the readable logic page.
I added the pop up window style data to the html. I also implemented the storing and displaying of data the is referencing logic inside other bodies of logic (like canHellRun for example). I colored the logic references green so they will appear different and look interactable to the user.